(ingestion) Fix EOF issue when ingesting staking ledgers
A regression was introduced when add support for hard linked
files. Because we now listen for CREATE file signals, there is a
scenario where we try to parse the file before it's finished being
written to disk. Instead we should listen for the CLOSE event in the
case that the file is not a hard link.