Remove accursed interactive prompts in CI scripts
These interactive prompts cause long-running invocations of tier-3 CI
runs to have to be monitored because an interactive prompt halts their
execution mid-way. This is unexpected. The answers to the prompts cause
confusion about what was tested and what was not tested because what is
tested depends on the answer to the prompts, which is undesirable.